@@ -1,5 +1,5 @@
 {
-  "name": "debug-react",
+  "name": "debug-react19",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
@@ -36,6 +36,7 @@
     "file-loader": "6.1.1",
     "fre": "^2.0.7",
     "fs-extra": "^9.0.1",
+    "hermes-eslint": "^0.19.1",
     "html-webpack-plugin": "4.5.0",
     "identity-obj-proxy": "3.0.0",
     "jest": "26.6.0",
@@ -51,10 +52,11 @@
     "postcss-preset-env": "6.7.0",
     "postcss-safe-parser": "5.0.2",
     "prompts": "2.4.0",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-app-polyfill": "^2.0.0",
     "react-dev-utils": "^11.0.1",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
+    "react-error-boundary": "^4.0.13",
     "react-refresh": "^0.8.3",
     "resolve": "1.18.1",
     "resolve-url-loader": "^3.1.2",
@@ -71,7 +73,7 @@
     "workbox-webpack-plugin": "5.1.4"
   },
   "scripts": {
-    "start": "node scripts/start.js",
+    "start": "node --openssl-legacy-provider scripts/start.js",
     "build": "node scripts/build.js",
     "test": "node scripts/test.js"
   },
