@@ -23,25 +23,25 @@ export default class LifeCyclePage extends React.Component {
     console.log("constructor", this.state.count);
   }
 
-  static getDerivedStateFromProps(props, state) {
-    // getDerivedStateFromProps 会在调用 render 方法之前调用，
-    //并且在初始挂载及后续更新时都会被调用。
-    //它应返回一个对象来更新 state，如果返回 null 则不更新任何内容。
-    const {count} = state;
-    console.log("getDerivedStateFromProps", count);
-    return count < 5 ? null : {count: 0};
-  }
+  // static getDerivedStateFromProps(props, state) {
+  //   // getDerivedStateFromProps 会在调用 render 方法之前调用，
+  //   //并且在初始挂载及后续更新时都会被调用。
+  //   //它应返回一个对象来更新 state，如果返回 null 则不更新任何内容。
+  //   const { count } = state;
+  //   console.log("getDerivedStateFromProps", count);
+  //   return count < 5 ? null : { count: 0 };
+  // }
   //在render之后，在componentDidUpdate之前。
-  getSnapshotBeforeUpdate(prevProps, prevState, snapshot) {
-    const {count} = prevState;
-    console.log("getSnapshotBeforeUpdate", count);
-    // return null;
-    return {pos: {x: 100, y: 101}};
-  }
-  /* UNSAFE_componentWillMount() {
+  // getSnapshotBeforeUpdate(prevProps, prevState, snapshot) {
+  //   const { count } = prevState;
+  //   console.log("getSnapshotBeforeUpdate", count);
+  //   // return null;
+  //   return { pos: { x: 100, y: 101 } };
+  // }
+  UNSAFE_componentWillMount() {
     //不推荐，将会被废弃
     console.log("componentWillMount", this.state.count);
-  } */
+  }
   componentDidMount() {
     console.log("componentDidMount", this.state.count);
   }
@@ -49,16 +49,16 @@ export default class LifeCyclePage extends React.Component {
     //组件卸载之前
     console.log("componentWillUnmount", this.state.count);
   }
-  /* UNSAFE_componentWillUpdate() {
+  UNSAFE_componentWillUpdate() {
     //不推荐，将会被废弃
     console.log("componentWillUpdate", this.state.count);
-  } */
+  }
   componentDidUpdate(prevProps, prevState, snapshot) {
     console.log("componentDidUpdate", this.state.count, prevState, snapshot);
   }
 
   shouldComponentUpdate(nextProps, nextState) {
-    const {count} = nextState;
+    const { count } = nextState;
     console.log("shouldComponentUpdate", count, nextState.count);
     return count !== 3;
   }
@@ -70,7 +70,7 @@ export default class LifeCyclePage extends React.Component {
   };
 
   render() {
-    const {count} = this.state;
+    const { count } = this.state;
     console.log("render", this.state);
     return (
       <div>
@@ -96,7 +96,9 @@ class Child extends React.Component {
   }
   render() {
     return (
-      <div style={{border: "solid 1px black", margin: "10px", padding: "10px"}}>
+      <div
+        style={{ border: "solid 1px black", margin: "10px", padding: "10px" }}
+      >
         我是Foo组件
         <div>Foo count: {this.props.count}</div>
       </div>
