@@ -2,7 +2,11 @@ import Button from "../components/Button";
 import Num from "../components/Num";
 import User from "../components/User";
 import { fetchData } from "../utils";
-import { Suspense, SuspenseList, useState } from "../whichReact";
+import {
+  Suspense,
+  // SuspenseList,
+  useState,
+} from "../whichReact";
 import ErrorBoundaryPage from "./ErrorBoundaryPage";
 
 const initialResource = fetchData();
@@ -13,17 +17,17 @@ export default function TransitionPage(props) {
     <div>
       <h3>TransitionPage</h3>
 
-      <ErrorBoundaryPage fallback={<h1>出错了</h1>}>
-        <SuspenseList revealOrder="forwards" tail="collapsed">
-          <Suspense fallback={<h1>Loading user..</h1>}>
-            <User resource={resource} />
-          </Suspense>
+      {/* <ErrorBoundaryPage fallback={<h1>出错了</h1>}> */}
+      {/* <SuspenseList revealOrder="forwards" tail="collapsed"> */}
+      <Suspense fallback={<h1>Loading user..</h1>}>
+        <User resource={resource} />
+      </Suspense>
 
-          <Suspense fallback={<h1>Loading num..</h1>}>
-            <Num resource={resource} />
-          </Suspense>
-        </SuspenseList>
-      </ErrorBoundaryPage>
+      <Suspense fallback={<h1>Loading num..</h1>}>
+        <Num resource={resource} />
+      </Suspense>
+      {/* </SuspenseList> */}
+      {/* </ErrorBoundaryPage> */}
 
       {/* <button onClick={() => setresource(fetchData())}>refresh</button> */}
 
