@@ -1,5 +1,4 @@
-import * as React from "react";
-import {useState, useCallback, PureComponent} from "react";
+import { useState, useCallback, memo, useEffect } from "react";
 
 export default function UseCallbackPage(props) {
   const [count, setCount] = useState(0);
@@ -25,20 +24,34 @@ export default function UseCallbackPage(props) {
       <p>{count}</p>
       <button onClick={() => setCount(count + 1)}>add</button>
       <input value={value} onChange={(event) => setValue(event.target.value)} />
-      <Child addClick={addClick} />
+      <ChildMemo addClick={addClick} />
     </div>
   );
 }
 
-class Child extends PureComponent {
-  render() {
-    console.log("child render");
-    const {addClick} = this.props;
-    return (
-      <div>
-        <h3>Child</h3>
-        <button onClick={() => console.log(addClick())}>add</button>
-      </div>
-    );
-  }
-}
+const ChildMemo = memo(function Child({ addClick }) {
+  useEffect(() => {
+    return () => {
+      console.log("destroy"); //sy-log
+    };
+  }, []);
+  console.log("Child"); //sy-log
+  return (
+    <div className="border">
+      <button onClick={() => console.log(addClick())}>add</button>
+    </div>
+  );
+});
+
+// class Child extends PureComponent {
+//   render() {
+//     console.log("child render");
+//     const { addClick } = this.props;
+//     return (
+//       <div>
+//         <h3>Child</h3>
+//         <button onClick={() => console.log(addClick())}>add</button>
+//       </div>
+//     );
+//   }
+// }
